Add files via upload
Configuring SSL (Secure Sockets Layer) virtual private network (VPN). The purpose 
is for some employee to have access to organization data remotely without the 
appearance on premises. Configuring SSL VPN will give them secure access to 
access the organization server. SSL will verify and encrypt the data between the 
clients and the servers, it is use to have provide high security communication 
between web brewers, server and the client. 
This LAB will be done using two department server, a firewall, two switches and 
two employee devices to connect remotely. The firewall (5505) will be placed 
between the two switches (2960), one switch will connect to the server and the 
second switch will connect to the remote users. The switch that will connect 
between firewall and server will be the incoming switch while the second switch 
will be the outgoing and will connect the employee devices to the firewall. IP will 
be given to the servers and devices statically.